refactor for tidy and neat signal handling
